+++ lib/src/store_flowable_factory.dart
@@ -1,5 +1,5 @@
 import 'package:store_flowable/src/base_store_flowable_factory.dart';
 
-abstract class StoreFlowableFactory<KEY, DATA> implements BaseStoreFlowableFactory<KEY, DATA> {
-  Future<DATA> fetchDataFromOrigin();
+abstract class StoreFlowableFactory<KEY, DATA, PARAM> implements BaseStoreFlowableFactory<KEY, DATA, PARAM> {
+  Future<DATA> fetchDataFromOrigin(final PARAM param);
 }